timestamps display fixed

+++ srcs/timestamps/get_time_ms.c
@@ -1,6 +1,6 @@
 #include "ft_ping.h"
 
-double	get_time_ms(struct timeval tv)
+float	get_time_ms(struct timeval tv)
 {
-	return tv.tv_sec * 1000 + tv.tv_usec / 1000;
+	return tv.tv_sec * 1000 + tv.tv_usec / 1000.0;
 }
